requirements.txt
django
python-decouple
psycopg2-binary
djangorestframework
djangorestframework_simplejwt
django-cors-headers
markdown
django-filter
drf-yasg
pylint